+++ tests/fixture/tg2app.py
@@ -111,3 +111,8 @@ class BasicTGController(TGController):
     def get_parameter(self, something):
         # Checking that parameters are received
         return 'Parameter received: %s' % something
+    
+    @expose()
+    def boolean_predicate(self):
+        p = not_anonymous()
+        return 'The predicate is %s' % bool(p)